+++ app/src/main/java/com/codesoom/assignment/dto/session/SessionRequestData.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dto;
+package com.codesoom.assignment.dto.session;
 
 import lombok.Getter;
