+++ app/src/main/java/com/codesoom/assignment/dto/user/UserResultData.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dto;
+package com.codesoom.assignment.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
